AP_HAL: AnalogIn uses only sized int types

+++ libraries/AP_HAL/AnalogIn.h
@@ -14,8 +14,8 @@ public:
 class AP_HAL::AnalogIn {
 public:
     virtual void init(void* implspecific) = 0;
-    virtual AP_HAL::AnalogSource* channel(int n) = 0;
-    virtual AP_HAL::AnalogSource* channel(int n, float scale) = 0;
+    virtual AP_HAL::AnalogSource* channel(int16_t n) = 0;
+    virtual AP_HAL::AnalogSource* channel(int16_t n, float scale) = 0;
 };
 
 #define ANALOG_INPUT_BOARD_VCC 254
